+++ R/as-nlist.R
@@ -6,15 +6,24 @@
 #' @return An nlist object.
 #' @export
 #' @examples
-#' as.nlist(list(x = 1:4))
-#' as.nlist(c(`a[2]` = 3, `a[1]` = 2))
-as.nlist <- function(x, ...) {
-  UseMethod("as.nlist")
+#' as_nlist(list(x = 1:4))
+#' as_nlist(c(`a[2]` = 3, `a[1]` = 2))
+as_nlist <- function(x, ...) {
+  UseMethod("as_nlist")
 }
 
-#' @describeIn as.nlist Coerce named numeric vector to nlist
+#' @rdname as_nlist
 #' @export
-as.nlist.numeric <- function(x, ...) {
+as.nlist <- function(x, ...) {
+  deprecate_soft("0.1.1", 
+                 what = "nlist::as.nlist()", 
+                 with = "nlist::as_nlist()")
+  UseMethod("as_nlist")
+}
+
+#' @describeIn as_nlist Coerce named numeric vector to nlist
+#' @export
+as_nlist.numeric <- function(x, ...) {
   chk_named(x)
   chk_term(as.term(names(x)), validate = "consistent", x_name = "`names(x)`")
   chk_not_any_na(names(x))
@@ -35,12 +44,12 @@ as.nlist.numeric <- function(x, ...) {
   x <- split(x, pars_terms(terms))
   x <- lapply(x, function(x) x[order(as.term(names(x)))])
   x <- lapply(x, function(x) set_dim(x, pdims(as.term(names(x)))[[1]]))
-  as.nlist(x)
+  as_nlist(x)
 }
 
-#' @describeIn as.nlist Coerce list to nlist
+#' @describeIn as_nlist Coerce list to nlist
 #' @export
-as.nlist.list <- function(x, ...) {
+as_nlist.list <- function(x, ...) {
   chk_unused(...)
 
   if (!length(x)) {
@@ -52,9 +61,9 @@ as.nlist.list <- function(x, ...) {
   x
 }
 
-#' @describeIn as.nlist Coerce data.frame to nlist
+#' @describeIn as_nlist Coerce data.frame to nlist
 #' @export
-as.nlist.data.frame <- function(x, ...) as.nlist(as.list(x))
+as_nlist.data.frame <- function(x, ...) as_nlist(as.list(x))
 
 #' @export
-as.nlist.nlist <- function(x, ...) x
+as_nlist.nlist <- function(x, ...) x