+++ R/as-nlists.R
@@ -6,27 +6,27 @@
 #' @return An nlists object.
 #' @export
 #' @examples
-#' as.nlists(list(nlist(x = c(1, 5)), nlist(x = c(2, 3)), nlist(x = c(3, 2))))
-as.nlists <- function(x, ...) {
-  UseMethod("as.nlists")
+#' as_nlists(list(nlist(x = c(1, 5)), nlist(x = c(2, 3)), nlist(x = c(3, 2))))
+as_nlists <- function(x, ...) {
+  UseMethod("as_nlists")
 }
 
-#' @describeIn as.nlists Coerce list to nlists
+#' @describeIn as_nlists Coerce list to nlists
 #' @export
-as.nlists.list <- function(x, ...) {
-  x <- lapply(x, as.nlist)
+as_nlists.list <- function(x, ...) {
+  x <- lapply(x, as_nlist)
   class(x) <- "nlists"
   chk_nlists(x)
   x
 }
 
-#' @describeIn as.nlists Coerce nlist to nlists
+#' @describeIn as_nlists Coerce nlist to nlists
 #' @export
-as.nlists.nlist <- function(x, ...) {
+as_nlists.nlist <- function(x, ...) {
   x <- list(x)
   class(x) <- "nlists"
   x
 }
 
 #' @export
-as.nlists.nlists <- function(x, ...) x
+as_nlists.nlists <- function(x, ...) x