+++ tests/testthat/test-as-nlists.R
@@ -1,14 +1,14 @@
-test_that("as.nlists.nlists", {
-  expect_identical(as.nlists(nlists()), nlists())
+test_that("as_nlists.nlists", {
+  expect_identical(as_nlists(nlists()), nlists())
 })
 
-test_that("as.nlists.nlist", {
+test_that("as_nlists.nlist", {
   expect_identical(
-    as.nlists(nlist()),
+    as_nlists(nlist()),
     structure(list(structure(list(), .Names = character(0), class = "nlist")), class = "nlists")
   )
   expect_identical(
-    as.nlists(nlist(x = 1)),
+    as_nlists(nlist(x = 1)),
     structure(list(structure(list(x = 1), class = "nlist")), class = "nlists")
   )
 })